Use double/cgfloat everywhere for vertical and horizontal  spacing. Because we compute line height/char width using ceil, using float for some calculations can result in a 1pt error. When using Menlo Regular 25pt with 80% hspacing and 80% vspacing, tmux tabs compute a height that is wrong because .8 is different enough between float and double that it computes 24 instead of 25 for the width (and likewise is off by 1 for height). Hopefully this fixes issue 9195

+++ sources/PTYSession.h
@@ -730,8 +730,8 @@ backgroundColor:(NSColor *)backgroundColor;
 - (void)changeFontSizeDirection:(int)dir;
 - (void)setFont:(NSFont*)font
     nonAsciiFont:(NSFont*)nonAsciiFont
-    horizontalSpacing:(float)horizontalSpacing
-    verticalSpacing:(float)verticalSpacing;
+    horizontalSpacing:(CGFloat)horizontalSpacing
+    verticalSpacing:(CGFloat)verticalSpacing;
 
 // Assigns a new GUID to the session so that changes to the bookmark will not
 // affect it. Returns the GUID of a divorced bookmark. Does nothing if already
